requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
beanie==1.26.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.14
dnspython==2.6.1
docarray==0.40.0
docstring_parser==0.16
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-cloud-trace==1.13.5
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib_metadata==8.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.25
langchain-experimental==0.0.63
langchain-google-vertexai==1.0.8
langchain-text-splitters==0.2.2
langsmith==0.1.94
lazy-model==0.2.0
markdown-it-py==3.0.0
marshmallow==3.21.3
mdurl==0.1.2
motor==3.5.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openinference-instrumentation==0.1.9
openinference-instrumentation-langchain==0.1.23
openinference-semantic-conventions==0.1.9
opentelemetry-api==1.26.0
opentelemetry-exporter-gcp-trace==1.6.0
opentelemetry-instrumentation==0.47b0
opentelemetry-resourcedetector-gcp==1.6.0a0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.6
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
setuptools==72.1.0
shapely==2.0.5
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
tenacity==8.3.0
toml==0.10.2
tqdm==4.66.4
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
